backend/python-excel/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
openpyxl==3.1.5
google-generativeai==0.8.3
python-dotenv==1.0.1
